Refactor HubController and IISController to use BaseGatewayController.
